pyproject.toml
[tool.poetry]
name = "medium-markdown-posts-tool"
version = "0.1.0"
description = ""
authors = ["liuliu <liuliugit@gmail.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.8"
requests = "^2.31.0"
python-dotenv = "^1.0.0"
click = "^8.1.7"

[tool.poetry.scripts]
medium = "src.medium.__main__:commands"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
